Change page name trimming to not rely on fwd slash
Previously this was hard coded to rely on fwd slash to properly trim the
page path string. This didn't work on windows because the page path was
using forward slash in the path name.

We instead can just rely on string slicing to trim the string based on
an index.